useを組み立てる部分を切り出す
