+++ src/autoUse.ts
@@ -1,60 +1,19 @@
 import { AutoUseRegex } from './autoUseRegex';
-import { DB } from './db';
-import { Selector } from './selector';
+import { UseBuilder } from './useBuilder';
 
-export class AutoUse {
-    private selector: Selector;
-
-    constructor(private _selector: Selector) {
-        this.selector = _selector;
-    }
-
-    private useBuilder(packageName: string, subList: string[] | undefined): string {
-        if (subList === undefined) { return 'use ' + packageName + ';'; }
-
-        let subListStr = "";
-        for (let i = 0; i < subList.length; i++) {
-            subListStr += (i !== subList.length - 1 ? subList[i] + ' ' : subList[i]);
-        }
-        return 'use ' + packageName + ' qw(' + subListStr + ');';
-    }
-
-    private async insertUseStatementByNames(names: string[]): Promise<boolean> {
-        let useStatements: string[] = [];
-        const declaredModuleSub = this.selector.getDeclaredModuleSub();
-
-        for (const name of names) {
-            const importObjects = DB.findByName(name);
-
-            if (importObjects.length > 0) {
-                const packageName = importObjects[0].packageName;
-                const alreadyDeclaredModuleSub = declaredModuleSub?.filter(dus => dus.packageName === packageName);
-                const alreadyDeclaredSubList = alreadyDeclaredModuleSub ? alreadyDeclaredModuleSub[0].subList : [];
-                const subList = alreadyDeclaredSubList.length > 0
-                    ? (alreadyDeclaredSubList.includes(name)
-                        ? alreadyDeclaredSubList
-                        : [...alreadyDeclaredSubList, name])
-                    : [name];
-                const useStatement = this.useBuilder(packageName, subList);
-                if (alreadyDeclaredModuleSub !== undefined) {
-                    const regex = `use ${packageName} qw(\\/|\\()(\\s*\\w+\\s*)*(\\/|\\));\n|\r\n|\r`;
-                    await this.selector.deleteByRegex(RegExp(regex, 'g'));
-                }
-                if (useStatement !== '') {
-                    useStatements.push(useStatement);
-                }
-            }
-        }
-        return this.selector.insertUseStatements(useStatements);
-    }
-
+export class AutoUse extends UseBuilder {
     private async insertFullyQualifiedModule(): Promise<boolean> {
         const declaredUse = this.selector.getDeclaredModule();
         const fullyQualifiedModules = this.selector.getFullyQualifiedModules();
-        const notDeclaredModule = declaredUse === undefined ? fullyQualifiedModules : fullyQualifiedModules?.filter(fqm => !declaredUse?.includes(fqm));
-        const useStatements = notDeclaredModule?.map(us => this.useBuilder(us, undefined));
+
+        const notDeclaredModule = declaredUse === undefined
+            ? fullyQualifiedModules
+            : fullyQualifiedModules?.filter(fqm => !declaredUse?.includes(fqm));
+
+        const useStatements = notDeclaredModule?.map(us => this.buildUseStatement(us, undefined));
 
         if (useStatements === undefined) { return Promise.reject('some error'); }
+
         return this.selector.insertUseStatements(useStatements);
     }
